So I've updated the program with the DLC now included =O I guess it's kind of DLC because you do download it in a way when you load the webpage. Anyways, I've implemented the Night Mode thingy, which activates when you press the shift button (3) as a command. Night mode supports its own map by just implementing some small changes that really boil down to me extending the size of the array to 30x30 instead of 15x15. I did this because I was thinking about "performance" (quotations cause it's a command line game, lol), and so I thought that we can just keep our operation times (except for shift as it has to shift night and regular mode consecutively making total operation N x 2 if N is 15 which tbh it kind of is cause either night mode or regular mode is 15x15.
All the features work in night mode too (TNT basically) and this is because each function basically just takes night mode now and will change
its iterators based on whether night mode is on or not, so for example, it accesses the backhalf of the array if night mode is on and the front half
if not. I thought this was kind of neat and adheres to "performance" optimisation.

I'm going to upload a fun test case that produces a larger field that is mirrored in the night realm, both fields include TNT so you can have fun blowing
that up :)

NIGHT REALM!